fix issue with beats not properly playing files imported with a different sample rate from bespoke's, and allow multiple files to be dropped at once onto beats slots

+++ Source/Sample.h
@@ -37,7 +37,7 @@ public:
    int NumChannels() const { return mData.NumActiveChannels(); }
    ChannelBuffer* Data() { return &mData; }
    int GetPlayPosition() const { return mOffset; }
-   void SetPlayPosition(int sample) { mOffset = sample; }
+   void SetPlayPosition(double sample) { mOffset = sample; }
    float GetSampleRateRatio() const { return mSampleRateRatio; }
    void Reset() { mOffset = mNumSamples; }
    void SetStopPoint(int stopPoint) { mStopPoint = stopPoint; }
